+++ config/namespaced/databases/config.go
@@ -23,12 +23,6 @@ func Configure(p *config.Provider) {
 			TerraformName: "ovh_cloud_project_database",
 		}
 	})
-	p.AddResourceConfigurator("ovh_cloud_project_database_ip_restriction", func(r *config.Resource) {
-		r.ShortGroup = shortGroup
-		r.References["cluster_id"] = config.Reference{
-			TerraformName: "ovh_cloud_project_database",
-		}
-	})
 	p.AddResourceConfigurator("ovh_cloud_project_database_kafka_acl", func(r *config.Resource) {
 		r.ShortGroup = shortGroup
 		r.References["cluster_id"] = config.Reference{
@@ -42,18 +36,6 @@ func Configure(p *config.Provider) {
 		}
 	})
 	p.AddResourceConfigurator("ovh_cloud_project_database_kafka_topic", func(r *config.Resource) {
-		r.ShortGroup = shortGroup
-		r.References["cluster_id"] = config.Reference{
-			TerraformName: "ovh_cloud_project_database",
-		}
-	})
-	p.AddResourceConfigurator("ovh_cloud_project_database_m3db_namespace", func(r *config.Resource) {
-		r.ShortGroup = shortGroup
-		r.References["cluster_id"] = config.Reference{
-			TerraformName: "ovh_cloud_project_database",
-		}
-	})
-	p.AddResourceConfigurator("ovh_cloud_project_database_m3db_user", func(r *config.Resource) {
 		r.ShortGroup = shortGroup
 		r.References["cluster_id"] = config.Reference{
 			TerraformName: "ovh_cloud_project_database",
@@ -84,12 +66,6 @@ func Configure(p *config.Provider) {
 		}
 	})
 	p.AddResourceConfigurator("ovh_cloud_project_database_postgresql_connection_pool", func(r *config.Resource) {
-		r.ShortGroup = shortGroup
-		r.References["cluster_id"] = config.Reference{
-			TerraformName: "ovh_cloud_project_database",
-		}
-	})
-	p.AddResourceConfigurator("ovh_cloud_project_database_redis_user", func(r *config.Resource) {
 		r.ShortGroup = shortGroup
 		r.References["cluster_id"] = config.Reference{
 			TerraformName: "ovh_cloud_project_database",